tests(refactor): Extract out individual ast component fixtures from json unit tests to conftest.
[ACTCOMP-36] This will stage our unit tests to use statically created ast nodes.
Also created a simple wrapper to register new nodes which can be more easily used in other modules.
